requirements.txt
# Frozen Pipe Analyzer Requirements

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.31.0

# Optional: For advanced analysis and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0

# Optional: For scheduled data collection
schedule>=1.2.0
